requirements_test.txt
pytest-django==2.8.0
django-nose==1.4.3
